Fix AI multicam exploits & a few other bugs

+++ code/modules/overmap/spacetravel.dm
@@ -63,10 +63,15 @@ proc/overmap_spacetravel(var/turf/space/T, var/atom/movable/A)
 	if (!M)
 		return
 
+	// Don't let AI eyes yeet themselves off the map
+	if(istype(A, /mob/observer/eye))
+		return
+
 	if(A.lost_in_space())
 		if(!QDELETED(A))
 			qdel(A)
 		return
+
 
 	var/nx = 1
 	var/ny = 1
